requirements.txt
Django==1.4.3
Fabric==1.5.2
Markdown==2.2.1
South==0.7.6
Werkzeug==0.8.3
bleach==1.1.5
django-bleach==0.1.3
django-compress==1.0.1
django-debug-toolbar==0.9.4
django-extensions==1.0.2
django-taggit==0.9.3
html5lib==0.95
lxml==3.1beta1
paramiko==1.9.0
pycrypto==2.6
wsgiref==0.1.2
